backend/requirements.txt
# Core
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.7
pydantic==2.5.2
python-dotenv==1.0.0

# Xử lý tài liệu
PyPDF2==3.0.1
python-docx==1.0.1
openpyxl==3.1.2
pdfplumber==0.10.3

# Xử lý AI và NLP
transformers==4.36.2
sentence-transformers==2.2.2
torch==2.1.2
spacy==3.7.2

# Web Crawling và Scraping
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.16.0
playwright==1.40.0
duckduckgo-search==4.1.0

# Tối ưu hóa
onnxruntime==1.16.3
accelerate==0.25.0

# Cơ sở dữ liệu
sqlalchemy==2.0.23
alembic==1.13.0